+++ tests/Types/FamilyTest.php
@@ -15,17 +15,25 @@ class FamilyTest extends TestCase
         $family = new Family([
             Family::ID   => 'id',
             Family::NAME => 'name',
+            Family::CLASSIFICATION => 'SAAS',
+            Family::ARROW_SUB_CATEGORIES => ['nce', 'ibm']
         ]);
 
         self::assertSame('id', $family->getId());
         self::assertSame('name', $family->getName());
+        self::assertSame('SAAS', $family->getClassification());
+        self::assertSame(['nce', 'ibm'], $family->getArrowSubCategories());
 
         $family
             ->setId('my id')
             ->setName('my name')
+            ->setClassification('my classification')
+            ->setArrowSubCategories(['first', 'second'])
         ;
 
         self::assertSame('my id', $family->getId());
         self::assertSame('my name', $family->getName());
+        self::assertSame('my classification', $family->getClassification());
+        self::assertSame(['first', 'second'], $family->getArrowSubCategories());
     }
 }