+++ CHANGELOG.md
@@ -1,9 +1,45 @@
-# Changelog
+  # Changelog
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+
+- `CatalogGraphQLClient` class:
+  - New method `findPriceBands()` to search for priceBands
+  - New method `findOnePriceBand()` to search for one PriceBand
+  - New method `findProducts()` to search for Products
+  - New method `findOneProduct()` to search for one Product
+
+- `Query` class:
+  - `GET_PRICE_BANDS` has been added to search for priceBands 
+  - `PRICE_BAND` has been added to search for one priceBand 
+
+- `PaginatedPriceBands.php` class is now available when searching for priceBands
+
+- `SearchBody` class:
+  - `GET_FAMILIES` has been added (boolean to set to `true` to get the families corresponding to the found priceBands)
+
+- `Family` class:
+  - `ARROW_SUB_CATEGORIES` has beed added (as an array of `string`)
+  - `CLASSIFICATION` has beed added (as a `string`)
+
+- `PriceBand` class:
+  - `OFFER` has been added (as a `OfferLight` class) 
+  - `PROGRAM` has been added (as a `Program` class) 
+  - `VENDOR` has been added (as a `Vendor` class) 
+
+- `OfferLight` class is now available inside `PriceBand` class when searching for priceBands
+
+### Changed
+- `CatalogGraphQLClient` class:
+  - The methods `find()` and `findOne()` are deprecated (replaced by `findProducts()` and `findOneProduct()`)
+
+### Fixed
+- `Family` class:
+  - @method Family setName(string `$id`) replaced by @method Family setName(string `$name`) 
 
 ## [0.6.13] - 2022-04-04